# Backend/main_api_spec.yml
openapi: 3.0.0
info:
  title: CMS API
  version: 1.0.0
  description: API for Content Management System

paths:
  /worker:
    post:
      summary: Create a new worker
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                manager:
                  type: string
                latitude:
                  type: number
                longitude:
                  type: number
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string

  /worker/{id}:
    get:
      summary: Get a worker by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  worker:
                    type: object
    delete:
      summary: Delete a worker by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  worker:
                    type: object

  /workers:
    get:
      summary: Get all workers
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Employee"

  /worker/status/{id}:
    post:
      summary: Update worker status
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  worker:
                    type: object

  /client:
    post:
      summary: Create a new client
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                desc:
                  type: string
                flag_18:
                  type: boolean
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string

  /client/{id}:
    get:
      summary: Get a client by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  client:
                    type: object
    delete:
      summary: Delete a client by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  client:
                    type: object

  /clients:
    get:
      summary: Get all clients
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  clients:
                    type: array
                    items:
                      type: object

  /admin:
    post:
      summary: Create a new admin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string

  /admin/{id}:
    get:
      summary: Get an admin by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  admin:
                    type: object
    delete:
      summary: Delete an admin by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  admin:
                    type: object

  /admins:
    get:
      summary: Get all admins
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  admin:
                    type: array
                    items:
                      type: object

  /location:
    post:
      summary: Create a new location
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                latitude:
                  type: number
                longitude:
                  type: number
                location_type:
                  type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string

  /locations:
    get:
      summary: Get all locations
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  locations:
                    type: array
                    items:
                      type: object

  /location/{userName}: # {70.093213,31.2121}
    get:
      summary: Get Latest workers Location
      parameters:
        - in: path
          name: UserName
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  latitude:
                    type: integer
                    example: 40.0888
                  longitude:
                    type: integer
                    example: -78.987
    delete:
      summary: Delete a location by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  location:
                    type: object

  /appointment:
    post:
      summary: Create a new appointment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                client:
                  type: string
                worker:
                  type: string
                start:
                  type: string
                end:
                  type: string
                location:
                  type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
    get:
      summary: Get all appointments
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  appointments:
                    type: array
                    items:
                      type: object

  /appointment/{id}:
    get:
      summary: Get an appointment by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  appointment:
                    type: object
    delete:
      summary: Delete an appointment by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  appointment:
                    type: object

components:
  schemas:
    Employee:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        client_name:
          type: string
        location:
          type: string
        status:
          type: string
        target_location:
          type: object
          properties:
            latitude:
              type: number
            longitude:
              type: number
        current_location:
          type: object
          properties:
            latitude:
              type: number
            longitude:
              type: number
        latest_update:
          type: string
          format: date-time
        check_in:
          type: string
          format: date-time
        checkout:
          type: string
          format: date-time